requirements.txt
clip @ git+https://github.com/openai/CLIP.git@main
matplotlib
numpy
opencv-python
pandas
Pillow
requests
scikit-learn
seaborn
torch
torchvision
tqdm
